requirements.txt
discord.py
python-dotenv
rcon
aiosqlite
